Add files via upload
Due to its geographical location, the Philippines is one of the most vulnerable countries globally when it comes to disasters, with flood being a recurrent issue. Disasters from flood hazards cause significant damages to livelihoods and human lives, with Marikina City among the most vulnerable areas being situated in the Marikina Valley. The study leveraged Power BI as a tool to conduct a detailed analysis of the impact of weather events from 2008 to 2023 to assess barangay-level vulnerability to flooding. Results revealed that areas with higher population densities and lower elevation, such as Tumana and Malanday, are the most flood-prone, with water levels reaching more than 3 meters during severe weather events. In contrast, barangays Marikina Heights and Parang are less susceptible to the hazard because of their high elevation. Meteorological data, population density, elevation are among the important factors in enhancing the disaster risk reduction and management strategies in Marikina City.